requirements.txt
Flask>=2.2.2
Werkzeug==2.2.2
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
psycopg2-binary>=2.9.1
captcha==0.5.0
